Let's lower the monster by half a pixel

+++ Popcorn/Monster.cpp
@@ -223,7 +223,7 @@ bool AMonster::Is_Finished()
 		return false;
 }
 //------------------------------------------------------------------------------------------------------------
-void AMonster::Activate(int x_pos, int y_pos, bool moving_right)
+void AMonster::Activate(double x_pos, double y_pos, bool moving_right)
 {
 	int random_speed;
 	int emitting_offset;
